Finished episode #14. Had some trouble with player update function.
Because I followed an extra video between episode #10 and #11 it
introduced a bug which stopped the charactrs from moving after a
collision. Therefore I reverted the update functuon in Player back to
what it was before these additions and then added episode #14's
additions to it. Seems to be working fine now.